Add protection for disabled FEDS when repacking
and moved WarningSummary (print only the first warning by default
and a summary at endStream) out of SiStripRawToDigiUnpacker for
reuse in SiStripDigiToRaw
