+++ EventFilter/SiStripRawToDigi/plugins/SiStripDigiToRawModule.h
@@ -37,6 +37,7 @@ namespace sistrip {
   
     void produce( edm::Event&, const edm::EventSetup& ) override;
     static void fillDescriptions(edm::ConfigurationDescriptions & descriptions);
+    void endStream() override;
   
   private:
